+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
+## 2025.03.25
+
+### Added
+
+- Designed abstract class for Playlist component
+
+### Updated
+
+- Changed design to include the code for all enhanced methods using the kernels
 
 ## 2025.02.03